Version includes catch for topoengine failure and repairs geometry when it occurs If virmem error occurs 10 times a popup window appears (if clicked enters infinite loop, though) Changed litter size and SD from integer to double Logs out memory use of process

+++ Ben_Branch/frmInput.cs
@@ -2845,8 +2845,8 @@ namespace SEARCH
                   System.Convert.ToDouble(this.txtResDieBetweenSeason.Text),
                   System.Convert.ToDouble(this.txtResBreedPercent.Text),
                   System.Convert.ToDouble(this.txtResFemalePercent.Text),
-                  System.Convert.ToInt32(this.txtResOffspringMean.Text),
-                  System.Convert.ToInt32(this.txtResOffspringSD.Text));
+                  System.Convert.ToDouble(this.txtResOffspringMean.Text),
+                  System.Convert.ToDouble(this.txtResOffspringSD.Text));
